backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
supabase>=2.3.0
asyncpg>=0.29.0

# Security
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Platform SDKs
google-ads>=24.0.0
facebook-business>=19.0.0

# AI
openai>=1.10.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2024.1

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.0
